implement word vector based search for future cluster creation.
