+++ chatterbot/comparisons/comparator.py
@@ -1,0 +1,7 @@
+class Comparator:
+
+    def __call__(self, statement_a, statement_b):
+        return self.compare(statement_a, statement_b)
+
+    def compare(self, statement_a, statement_b):
+        return 0
