+++ chatterbot/comparisons/levelshtein_distance.py
@@ -1,0 +1,45 @@
+from .comparator import Comparator
+
+# Use python-Levenshtein if available
+try:
+    from Levenshtein.StringMatcher import StringMatcher as SequenceMatcher
+except ImportError:
+    from difflib import SequenceMatcher
+
+
+class LevenshteinDistance(Comparator):
+    """
+    Compare two statements based on the Levenshtein distance
+    of each statement's text.
+
+    For example, there is a 65% similarity between the statements
+    "where is the post office?" and "looking for the post office"
+    based on the Levenshtein distance algorithm.
+    """
+
+    def compare(self, statement, other_statement):
+        """
+        Compare the two input statements.
+
+        :return: The percent of similarity between the text of the statements.
+        :rtype: float
+        """
+
+        # Return 0 if either statement has a falsy text value
+        if not statement.text or not other_statement.text:
+            return 0
+
+        # Get the lowercase version of both strings
+        statement_text = str(statement.text.lower())
+        other_statement_text = str(other_statement.text.lower())
+
+        similarity = SequenceMatcher(
+            None,
+            statement_text,
+            other_statement_text
+        )
+
+        # Calculate a decimal percent of the similarity
+        percent = round(similarity.ratio(), 2)
+
+        return percent
