+++ requirements.txt
@@ -5,3 +5,9 @@ pymongo>=3.3,<4.0
 python-dateutil>=2.7,<2.8
 sqlalchemy>=1.2,<1.3
 pint>=0.8.1
+sklearn
+numpy
+scipy
+spacy
+pytz
+#en_core_web_lg
